skiller-simulator-navgraph: adapt to the moved aspect
The aspect is now in a different location and not in the
skiller_simulator namespace anymore.
